feat(modules): move to new modules structure
