+++ src/model/index.js
@@ -1,1 +1,1 @@
-export User from './User';
+export User from '../modules/user/UserModel';
