+++ src/modules/user/UserConnection.js
@@ -1,14 +1,10 @@
 // @flow
 
-import {
-  GraphQLInt,
-} from 'graphql';
+import { GraphQLInt } from 'graphql';
 
-import {
-  connectionDefinitions,
-} from 'graphql-relay';
+import { connectionDefinitions } from 'graphql-relay';
 
-import UserType from '../type/UserType';
+import UserType from './UserType';
 
 export default connectionDefinitions({
   name: 'User',
