+++ src/modules/user/UserType.js
@@ -1,16 +1,8 @@
 // @flow
 
-import {
-  GraphQLObjectType,
-  GraphQLString,
-  GraphQLBoolean,
-} from 'graphql';
-import {
-  globalIdField,
-} from 'graphql-relay';
-import {
-  NodeInterface,
-} from '../interface/NodeInterface';
+import { GraphQLObjectType, GraphQLString, GraphQLBoolean } from 'graphql';
+import { globalIdField } from 'graphql-relay';
+import { NodeInterface } from '../../interface/NodeInterface';
 
 export default new GraphQLObjectType({
   name: 'User',
